Update README.md
Searching Bibliography File 
A Bibliography file allows authors of publishable materials (journals, books, conferences, etc …) to compile and structure the list of sources that have been referenced when writing a published document. It is common practice to structure a bibliography file based on information such as author, year of publication, type of publication, and, title. 
Problem Statement 
A customer is looking for a search tool for bibliography files. The search tool is expected to have the functionalities listed in table 2. 
 students will develop a system that can read data that has been stored within text files and write new data to the file. It is assumed that the files to be processed contain entries of documents published online. These files, known as bibliography files, will contain entries that follow a particular structure. Four categories of documents have been considered: Journal, Conference, Website and Book. An example of a bibliography entry is provided in Figure 1.